Move "beach down" code into helper function
