+++ cmd/beach/cmd/helpers.go
@@ -26,6 +26,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/flownative/localbeach/pkg/beachsandbox"
 	"github.com/flownative/localbeach/pkg/exec"
 	"github.com/flownative/localbeach/pkg/path"
 	log "github.com/sirupsen/logrus"
@@ -227,4 +228,77 @@ func startLocalBeach() error {
 		}
 	}
 	return nil
+}
+
+func bringBeachDown() bool {
+	instanceRoots, err := findInstanceRoots()
+	if err != nil {
+		log.Fatal(err)
+		return false
+	}
+	for _, instanceRoot := range instanceRoots {
+		log.Info("Stopping instance in " + instanceRoot + "...")
+		sandbox, err := beachsandbox.GetSandbox(instanceRoot)
+		if err != nil && !errors.Is(err, beachsandbox.ErrNoFlowFound) {
+			log.Fatal(err)
+			return false
+		}
+		commandArgs := []string{"compose", "-f", sandbox.DockerComposeFilePath, "rm", "--force", "--stop", "-v"}
+		output, err := exec.RunCommand("docker", commandArgs)
+		if err != nil {
+			log.Fatal(output)
+			return false
+		}
+	}
+
+	log.Info("Stopping reverse proxy and database server ...")
+	commandArgs := []string{"compose", "-f", filepath.Join(path.Base, "docker-compose.yml"), "rm", "--force", "--stop", "-v"}
+	output, err := exec.RunCommand("docker", commandArgs)
+	if err != nil {
+		log.Fatal(output)
+		return false
+	}
+	return true
+}
+
+func findInstanceRoots() ([]string, error) {
+	var configurationFiles []string
+
+	output, err := exec.RunCommand("docker", []string{"ps", "-q", "--filter", "network=local_beach"})
+	if err != nil {
+		return nil, errors.New(output)
+	}
+	for _, line := range strings.Split(output, "\n") {
+		containerID := strings.TrimSpace(line)
+		if len(containerID) > 0 {
+			output, err := exec.RunCommand("docker", []string{"inspect", "-f", "{{index .Config.Labels \"com.docker.compose.project.config_files\"}}", containerID})
+			if err != nil {
+				return nil, errors.New(output)
+			}
+			projectDirectory := filepath.Dir(strings.TrimSpace(output))
+			if containsLocalBeachInstance(projectDirectory) {
+				configurationFiles = append(configurationFiles, projectDirectory)
+			}
+		}
+	}
+
+	return removeDuplicates(configurationFiles), nil
+}
+
+func containsLocalBeachInstance(path string) bool {
+	path = filepath.Join(path, ".localbeach.docker-compose.yaml")
+	_, err := os.Stat(path)
+	return !errors.Is(err, os.ErrNotExist)
+}
+
+func removeDuplicates(strSlice []string) []string {
+	allKeys := make(map[string]bool)
+	var list []string
+	for _, item := range strSlice {
+		if _, value := allKeys[item]; !value {
+			allKeys[item] = true
+			list = append(list, item)
+		}
+	}
+	return list
 }
